Add vault wrapper image with setcap magic
The ryodi/vault image just wraps the upstream Vault image with a call to
`setcap` the IPC_LOCK capability on the vault binary so that the whole
container can be run as root.

While we're in there, we also set the two SKIP_* env vars, SKIP_CHOWN
(since we bind-mount config and don't need to chown anything) and
SKIP_SETCAP (since we already did it in build).